fix(auction): self-invocation 문제 해결
